+++ mysql-test/suite/rpl/t/rpl_row_slave_skip_errors.opt
@@ -1,0 +1,2 @@
+--slave-skip-errors=1677
+
