+++ mysql-test/suite/rpl/t/rpl_row_slave_skip_errors.test
@@ -1,0 +1,56 @@
+# ==== Purpose ====
+#
+# Check that slave-skip-errors can skip ER_SLAVE_CONVERSION_FAILED.
+#
+# ==== Implementation ====
+# On slave, set slave_skip_errors=1677, so that slave skips ER_SLAVE_CONVERSION_FAILED 
+# reported during application of row based events.
+# On master, create a table t with two columns, a VARCHAR with a length of 25 and an INT.
+# On slave, increase the VARCHAR length to 255. For any row, the slave leaves the VARCHAR
+# field as the default value, while the INT is replicated correctly.
+#
+# References
+#   MDEV-7270: cannot slave_skip_errors for 1677
+
+--source include/have_binlog_format_row.inc
+--source include/master-slave.inc
+
+# On master create table t which contains a field named 'name' with length
+# varchar(25).
+CREATE TABLE t (name VARCHAR(25) DEFAULT NULL, age INT);
+--source include/sync_slave_sql_with_master.inc
+
+# On slave alter the name field length to varchar(255).
+ALTER TABLE t MODIFY name VARCHAR(255);
+
+connection master;
+INSERT INTO t VALUE ('Omar', 22);
+--source include/sync_slave_sql_with_master.inc
+--source include/check_slave_no_error.inc
+
+connection slave;
+ALTER TABLE t alter name set default "slave_default";
+
+connection master;
+INSERT INTO t VALUE ('Khokh', 22);
+--source include/sync_slave_sql_with_master.inc
+--source include/check_slave_no_error.inc
+
+--echo # The skipped conversions take the default value defined for the column.
+connection master;
+SELECT * FROM t;
+
+connection slave;
+SELECT * FROM t;
+
+connection master;
+SELECT COUNT(*) AS master_count FROM t;
+connection slave;
+SELECT COUNT(*) AS slave_count FROM t;
+
+
+--echo ==== Clean up ====
+connection master;
+DROP TABLE t;
+--source include/rpl_end.inc
+
