+++ sql/log_event_server.cc
@@ -328,20 +328,6 @@ inline int idempotent_error_code(int err_code)
   return (ret);
 }
 
-/**
-  Ignore error code specified on command line.
-*/
-
-inline int ignored_error_code(int err_code)
-{
-  if (use_slave_mask && bitmap_is_set(&slave_error_mask, err_code))
-  {
-    statistic_increment(slave_skipped_errors, LOCK_status);
-    return 1;
-  }
-  return err_code == ER_SLAVE_IGNORED_TABLE;
-}
-
 /*
   This function converts an engine's error to a server error.
