+++ sql/slave.h
@@ -135,6 +135,20 @@ extern ulonglong slave_skipped_errors;
 extern const char *relay_log_index;
 extern const char *relay_log_basename;
 
+/**
+  Ignore error code specified on command line.
+*/
+
+inline int ignored_error_code(int err_code)
+{
+  if (use_slave_mask && bitmap_is_set(&slave_error_mask, err_code))
+  {
+    statistic_increment(slave_skipped_errors, LOCK_status);
+    return 1;
+  }
+  return err_code == ER_SLAVE_IGNORED_TABLE;
+}
+
 /*
   4 possible values for Master_info::slave_running and
   Relay_log_info::slave_running.
